refactor(http): keep management lifecycle below line gate
